Move dummy storage to helpers module, ensure views tests do not store dummy images
